Lager noen funksjoner om til moduler.
